hugeutils: allow for long mount entries in find_mounts
The buffer used for parsing the mount list is not large enough.

In container environments using overlay filesystems, the use of a
large number of layers can easily result in valid entries which
exceed the buffer size resulting in 'Line too long when parsing mounts'
errors and the mount list not being processed completely.

Signed-off-by: Thomas Abraham <tabraham@suse.com>